modify the hook of ssrf
