+++ agent/java/src/main/java/com/fuxi/javaagent/hook/sql/SQLDriverManagerHook.java
@@ -28,27 +28,27 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.fuxi.javaagent.hook;
+package com.fuxi.javaagent.hook.sql;
 
 import com.fuxi.javaagent.HookHandler;
+import com.fuxi.javaagent.hook.AbstractClassHook;
 import com.fuxi.javaagent.plugin.checker.CheckParameter;
-import com.fuxi.javaagent.plugin.js.engine.JSContext;
-import com.fuxi.javaagent.plugin.js.engine.JSContextFactory;
-import org.mozilla.javascript.Scriptable;
+import com.fuxi.javaagent.plugin.checker.policy.SqlConnectionChecker;
+import com.fuxi.javaagent.tool.TimeUtils;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.Type;
 import org.objectweb.asm.commons.AdviceAdapter;
 import org.objectweb.asm.commons.Method;
 
-import java.io.File;
-import java.io.IOException;
+import java.util.HashMap;
+import java.util.Properties;
 
 /**
- * Created by zhuming01 on 5/16/17.
+ * Created by lxk on 7/6/17.
  * All rights reserved
  */
-public class FileHook extends AbstractClassHook {
+public class SQLDriverManagerHook extends AbstractClassHook {
     /**
      * (none-javadoc)
      *
@@ -56,7 +56,7 @@ public class FileHook extends AbstractClassHook {
      */
     @Override
     public String getType() {
-        return "directory";
+        return "sql";
     }
 
     /**
@@ -66,7 +66,7 @@ public class FileHook extends AbstractClassHook {
      */
     @Override
     public boolean isClassMatched(String className) {
-        return "java/io/File".equals(className);
+        return "java/sql/DriverManager".equals(className);
     }
 
     /**
@@ -75,14 +75,25 @@ public class FileHook extends AbstractClassHook {
      * @see com.fuxi.javaagent.hook.AbstractClassHook#hookMethod(int, String, String, String, String[], MethodVisitor)
      */
     @Override
-    public MethodVisitor hookMethod(int access, String name, String desc, String signature, String[] exceptions, MethodVisitor mv) {
-        if (name.equals("listFiles")) {
+    public MethodVisitor hookMethod(int access, String name, String desc,
+                                    String signature, String[] exceptions, MethodVisitor mv) {
+        if (name.equals("getConnection") && desc.startsWith("(Ljava/lang/String;Ljava/util/Properties;"
+                + "Ljava/lang/Class")) {
             return new AdviceAdapter(Opcodes.ASM5, mv, access, name, desc) {
                 @Override
                 protected void onMethodEnter() {
-                    loadThis();
-                    invokeStatic(Type.getType(FileHook.class),
-                            new Method("checkListFiles", "(Ljava/io/File;)V"));
+                    loadArg(0);
+                    loadArg(1);
+                    invokeStatic(Type.getType(SQLDriverManagerHook.class),
+                            new Method("checkSqlConnection", "(Ljava/lang/String;Ljava/util/Properties;)V"));
+                    invokeStatic(Type.getType(HookHandler.class),
+                            new Method("preShieldHook", "()V"));
+                }
+
+                @Override
+                protected void onMethodExit(int i) {
+                    invokeStatic(Type.getType(HookHandler.class),
+                            new Method("postShieldHook", "()V"));
                 }
             };
         }
@@ -90,23 +101,18 @@ public class FileHook extends AbstractClassHook {
     }
 
     /**
-     * 列出文件列表方法hook点
+     * 检测sql连接规范
      *
-     * @param file 文件对象
+     * @param url        连接url
+     * @param properties 连接属性
      */
-    public static void checkListFiles(File file) {
-        if (file != null) {
-            JSContext cx = JSContextFactory.enterAndInitContext();
-            Scriptable params = cx.newObject(cx.getScope());
-            params.put("path", params, file.getPath());
-            try {
-                params.put("realpath", params, file.getCanonicalPath());
-            } catch (IOException e) {
-                params.put("realpath", params, file.getAbsolutePath());
-            }
-
-            HookHandler.doCheck(CheckParameter.Type.DIRECTORY, params);
+    public static void checkSqlConnection(String url, Properties properties) {
+        Long lastAlarmTime = SqlConnectionChecker.alarmTimeCache.get(url);
+        if (lastAlarmTime == null || (System.currentTimeMillis() - lastAlarmTime) > TimeUtils.DAY_MILLISECOND) {
+            HashMap<String, Object> params = new HashMap<String, Object>(4);
+            params.put("url", url);
+            params.put("properties", properties);
+            HookHandler.doCheckWithoutRequest(CheckParameter.Type.POLICY_SQL_CONNECTION, params);
         }
     }
-
 }
