+++ agent/java/src/main/java/com/fuxi/javaagent/hook/ssrf/URLConnectionHook.java
@@ -1,0 +1,87 @@
+/*
+ * Copyright (c) 2017 Baidu, Inc. All Rights Reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice,
+ * this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ *
+ * 3. Neither the name of the copyright holder nor the names of its contributors
+ * may be used to endorse or promote products derived from this software without
+ * specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package com.fuxi.javaagent.hook.ssrf;
+
+import com.fuxi.javaagent.HookHandler;
+import org.objectweb.asm.MethodVisitor;
+import org.objectweb.asm.Opcodes;
+import org.objectweb.asm.Type;
+import org.objectweb.asm.commons.AdviceAdapter;
+import org.objectweb.asm.commons.Method;
+
+import java.net.URL;
+import java.net.URLConnection;
+
+/**
+ * Created by tyy on 17-12-7.
+ *
+ * jdk 中进行 http 请求的 hook 点
+ */
+public class URLConnectionHook extends AbstractSSRFHook {
+    @Override
+    public boolean isClassMatched(String className) {
+        return "sun/net/www/protocol/http/HttpURLConnection".equals(className);
+    }
+
+    @Override
+    public String getType() {
+        return "ssrf";
+    }
+
+    @Override
+    protected MethodVisitor hookMethod(int access, String name, String desc, String signature, String[] exceptions, MethodVisitor mv) {
+        if (name.equals("connect") && desc.equals("()V")) {
+            return new AdviceAdapter(Opcodes.ASM5, mv, access, name, desc) {
+                @Override
+                protected void onMethodEnter() {
+                    loadThis();
+                    invokeStatic(Type.getType(URLConnectionHook.class),
+                            new Method("checkHttpConnection", "(Ljava/net/URLConnection;)V"));
+                }
+            };
+        }
+        return mv;
+    }
+
+    public static void checkHttpConnection(URLConnection urlConnection) {
+        try {
+            if (urlConnection != null) {
+                URL url = urlConnection.getURL();
+                if (url != null) {
+                    checkHttpUrl(url.toString(), urlConnection.getURL().getHost());
+                }
+            }
+        } catch (Exception e) {
+            HookHandler.LOGGER.warn(e.getMessage());
+        }
+    }
+
+}
